further clean up `lib.rs`
